PET-103 refactor : Auth-Application Dto 패키지 구조 통일
