+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/dto/Provider.java
@@ -1,4 +1,4 @@
-package com.pawith.authapplication.service.dto;
+package com.pawith.authapplication.dto;
 
 public enum Provider {
     GOOGLE, NAVER, KAKAO,APPLE
