+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/OAuthUseCase.java
@@ -1,7 +1,7 @@
 package com.pawith.authapplication.service;
 
-import com.pawith.authapplication.service.dto.OAuthResponse;
-import com.pawith.authapplication.service.dto.Provider;
+import com.pawith.authapplication.dto.OAuthResponse;
+import com.pawith.authapplication.dto.Provider;
 
 public interface OAuthUseCase {
