+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/command/handler/AuthHandler.java
@@ -1,7 +1,7 @@
 package com.pawith.authapplication.service.command.handler;
 
-import com.pawith.authapplication.service.dto.OAuthRequest;
-import com.pawith.authapplication.service.dto.OAuthUserInfo;
+import com.pawith.authapplication.dto.OAuthRequest;
+import com.pawith.authapplication.dto.OAuthUserInfo;
 
 public interface AuthHandler{
     OAuthUserInfo handle(OAuthRequest authenticationInfo);
