+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/command/handler/impl/AppleOAuthHandler.java
@@ -3,9 +3,9 @@ package com.pawith.authapplication.service.command.handler.impl;
 import com.pawith.authapplication.service.command.feign.AppleFeignClient;
 import com.pawith.authapplication.service.command.feign.response.Keys;
 import com.pawith.authapplication.service.command.handler.AuthHandler;
-import com.pawith.authapplication.service.dto.OAuthRequest;
-import com.pawith.authapplication.service.dto.OAuthUserInfo;
-import com.pawith.authapplication.service.dto.Provider;
+import com.pawith.authapplication.dto.OAuthRequest;
+import com.pawith.authapplication.dto.OAuthUserInfo;
+import com.pawith.authapplication.dto.Provider;
 import com.pawith.commonmodule.exception.Error;
 import com.pawith.authdomain.jwt.exception.InvalidTokenException;
 import io.jsonwebtoken.*;
