+++ Domain-Module/Auth-Module/Auth-Presentation/src/test/java/com/pawith/authpresentation/OAuthControllerTest.java
@@ -3,8 +3,8 @@ package com.pawith.authpresentation;
 import com.navercorp.fixturemonkey.FixtureMonkey;
 import com.navercorp.fixturemonkey.api.introspector.ConstructorPropertiesArbitraryIntrospector;
 import com.pawith.authapplication.service.OAuthUseCase;
-import com.pawith.authapplication.service.dto.OAuthResponse;
-import com.pawith.authapplication.service.dto.Provider;
+import com.pawith.authapplication.dto.OAuthResponse;
+import com.pawith.authapplication.dto.Provider;
 import com.pawith.authpresentation.common.FilterConfig;
 import com.pawith.commonmodule.BaseRestDocsTest;
 import org.junit.jupiter.api.DisplayName;
